refactor: handCard의 타입 변환을 CardUnit에서 하도록 수정
players에서 changeToStand()시 검사하도록 수정
